requirements.txt
# Core Data Science
jupyterlab==4.0.10
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.4

# LLM and AI
langchain==0.1.0
langchain-community==0.0.10
langchain-experimental==0.0.47
langgraph==0.0.20
openai==1.6.1
tiktoken==0.5.2

# Embeddings and Vector Store
sentence-transformers==2.2.2
chromadb==0.4.22

# Monitoring
langfuse==2.13.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
httpx==0.26.0
tenacity==8.2.3

# SQL and Text2SQL
sqlparse==0.4.4

# Visualization
ipywidgets==8.1.1
